requirements.txt
python-binance==1.0.19
pandas==2.1.1
numpy==1.24.3
ta==0.10.2
python-dotenv==1.0.0
ccxt==4.4.44
PyQt6==6.5.0
pyqtgraph==0.13.3
